12/11

+++ Snake/Settings.cs
@@ -8,8 +8,8 @@ namespace Snake
 {
     internal class Settings
     {
-        public static int height { get; set; }
-        public static int width { get; set; }
+        public static float height { get; set; }
+        public static float width { get; set; }
 
         public static string directions;
